Unirse a tanda funcionando desde UI
